Matlab: Rename to matlab
This complies to our package naming conventions.
